+++ src/app/shared/app-demo-generators.ts
@@ -7,6 +7,7 @@ import {
   type HostingMenuItem
 } from './demo-data';
 import type {
+  ActivityDateTimeRange,
   ActivityListRow,
   ActivityMemberEntry,
   ActivityMemberStatus,
@@ -15,8 +16,14 @@ import type {
   AssetMemberRequest,
   AssetRequestStatus,
   AssetType,
+  EventCapacityRange,
+  EventFeedbackCard,
+  EventFeedbackOption,
   EventExploreCard,
-  ExperienceEntry
+  ExperienceEntry,
+  SubEventFormItem,
+  SubEventGroupItem,
+  SubEventsDisplayMode
 } from './app-types';
 import { AppUtils } from './app-utils';
 
@@ -278,6 +285,83 @@ export class AppDemoGenerators {
     };
   }
 
+  static buildEventFeedbackCards(options: {
+    eventItems: EventMenuItem[];
+    users: DemoUser[];
+    activeUser: DemoUser;
+    eventDatesById: Record<string, string>;
+    activityImageById: Record<string, string>;
+    eventFeedbackUnlockDelayMs: number;
+    eventOverallOptions: EventFeedbackOption[];
+    hostImproveOptions: EventFeedbackOption[];
+    attendeeCollabOptions: EventFeedbackOption[];
+    attendeeRejoinOptions: EventFeedbackOption[];
+  }): EventFeedbackCard[] {
+    const nowMs = Date.now();
+    const eventCards: EventFeedbackCard[] = [];
+    for (const item of options.eventItems) {
+      if (item.isAdmin) {
+        continue;
+      }
+      const startMs = this.eventStartAtMs(item.id, options.eventDatesById);
+      if (startMs === null || nowMs < startMs + options.eventFeedbackUnlockDelayMs) {
+        continue;
+      }
+      const eventLabel = this.eventFeedbackWhenLabel(item.id, options.eventDatesById);
+      const host = this.feedbackHostUserForEvent(item.id, options.users, options.activeUser);
+      const attendees = this.feedbackAttendeesForEvent(item.id, host.id, options.users, options.activeUser.id);
+      eventCards.push({
+        id: `feedback-event-${item.id}`,
+        eventId: item.id,
+        kind: 'event',
+        targetUserId: host.id,
+        targetRole: 'Admin',
+        icon: 'event_available',
+        imageUrl: options.activityImageById[item.id] ?? `https://picsum.photos/seed/event-feedback-card-${item.id}/1200/700`,
+        toneClass: 'feedback-card-tone-event feedback-role-admin',
+        heading: item.title,
+        subheading: `${eventLabel} · ${item.shortDescription}`,
+        identityTitle: `${host.name} · Host`,
+        identitySubtitle: `Admin · ${host.city}`,
+        identityStatusClass: 'member-status-admin',
+        identityStatusIcon: 'admin_panel_settings',
+        questionPrimary: `How did ${item.title} feel for you overall?`,
+        questionSecondary: `What should ${host.name} improve next time?`,
+        primaryOptions: options.eventOverallOptions,
+        secondaryOptions: options.hostImproveOptions,
+        answerPrimary: '',
+        answerSecondary: ''
+      });
+      for (const attendee of attendees) {
+        const attendeeRole = this.feedbackRoleForAttendee(item.id, attendee.id);
+        eventCards.push({
+          id: `feedback-attendee-${item.id}-${attendee.id}`,
+          eventId: item.id,
+          kind: 'attendee',
+          attendeeUserId: attendee.id,
+          targetUserId: attendee.id,
+          targetRole: attendeeRole,
+          icon: 'groups',
+          imageUrl: attendee.images?.[0] ?? `https://i.pravatar.cc/1200?img=${(this.hashText(`feedback-attendee:${item.id}:${attendee.id}`) % 70) + 1}`,
+          toneClass: `feedback-card-tone-attendee ${this.feedbackRoleToneClass(attendeeRole)}`,
+          heading: `${attendee.name} · ${item.title}`,
+          subheading: `Attendee feedback · ${eventLabel}`,
+          identityTitle: `${attendee.name}, ${attendee.age}`,
+          identitySubtitle: `${attendeeRole} · ${attendee.city}`,
+          identityStatusClass: this.feedbackRoleStatusClass(attendeeRole),
+          identityStatusIcon: this.feedbackRoleStatusIcon(attendeeRole),
+          questionPrimary: `How was collaboration with ${attendee.name} (${attendee.traitLabel}) during this event?`,
+          questionSecondary: `Would you team up with ${attendee.name} again in a future event?`,
+          primaryOptions: options.attendeeCollabOptions,
+          secondaryOptions: options.attendeeRejoinOptions,
+          answerPrimary: '',
+          answerSecondary: ''
+        });
+      }
+    }
+    return eventCards;
+  }
+
   static resolveSectionBadge(values: number[], itemCount: number): number {
     const positiveTotal = values.reduce((sum, value) => sum + (value > 0 ? value : 0), 0);
     if (positiveTotal > 0) {
@@ -333,6 +417,69 @@ export class AppDemoGenerators {
       }
     }
     return seeded;
+  }
+
+  static seededEventCapacityRange(
+    eventId: string,
+    activityCapacityById: Record<string, string>
+  ): EventCapacityRange {
+    const source = activityCapacityById[eventId];
+    if (source) {
+      const parts = source.split('/').map(part => Number.parseInt(part.trim(), 10));
+      if (parts.length >= 2 && Number.isFinite(parts[0]) && Number.isFinite(parts[1])) {
+        const min = Math.max(0, Math.min(parts[0], parts[1]));
+        const max = Math.max(min, parts[1]);
+        return { min, max };
+      }
+    }
+    const seed = this.hashText(`event-capacity:${eventId}`);
+    const max = 10 + (seed % 24);
+    const min = Math.max(0, Math.floor(max * 0.45));
+    return { min, max };
+  }
+
+  static buildSeededSubEventsForEvent(
+    source: EventMenuItem | HostingMenuItem,
+    options: {
+      isHosting: boolean;
+      activityDateTimeRangeById: Record<string, ActivityDateTimeRange>;
+      hostingDatesById: Record<string, string>;
+      eventDatesById: Record<string, string>;
+      eventCapacityById: Record<string, EventCapacityRange>;
+      activityCapacityById: Record<string, string>;
+      defaultStartIso: string;
+      activeUserId: string;
+    }
+  ): SubEventFormItem[] {
+    const dateSource = options.activityDateTimeRangeById[source.id];
+    const fallbackStartIso = options.isHosting
+      ? (options.hostingDatesById[source.id] ?? options.defaultStartIso)
+      : (options.eventDatesById[source.id] ?? options.defaultStartIso);
+    const start = new Date(dateSource?.startIso ?? fallbackStartIso);
+    const end = new Date(
+      dateSource?.endIso
+      ?? new Date(start.getTime() + (4 * 60 * 60 * 1000)).toISOString().slice(0, 19)
+    );
+    const startMs = Number.isNaN(start.getTime()) ? Date.now() : start.getTime();
+    const endMs = Number.isNaN(end.getTime()) || end.getTime() <= startMs
+      ? (startMs + (4 * 60 * 60 * 1000))
+      : end.getTime();
+    const seed = this.hashText(`event-subevents:${source.id}:${source.title}:${source.shortDescription}`);
+    const tournamentMode = (seed % 3) === 0;
+    const eventCapacity = options.eventCapacityById[source.id]
+      ?? this.seededEventCapacityRange(source.id, options.activityCapacityById);
+    const eventMax = this.normalizedEventCapacityValue(eventCapacity.max) ?? 0;
+    if (tournamentMode) {
+      return this.buildSeededTournamentSubEvents(source, startMs, endMs, seed, options.activeUserId, eventMax);
+    }
+    return this.buildSeededCasualSubEvents(source, startMs, endMs, seed, options.activeUserId, eventMax);
+  }
+
+  static inferredSubEventsDisplayMode(items: SubEventFormItem[]): SubEventsDisplayMode {
+    if (items.some(item => !item.optional && (item.groups?.length ?? 0) > 0)) {
+      return 'Tournament';
+    }
+    return 'Casual';
   }
 
   static isFriendOfActiveUser(userId: string, activeUserId: string): boolean {
@@ -510,5 +657,228 @@ export class AppDemoGenerators {
       status,
       note
     };
+  }
+
+  private static buildSeededCasualSubEvents(
+    source: EventMenuItem | HostingMenuItem,
+    startMs: number,
+    endMs: number,
+    seed: number,
+    activeUserId: string,
+    eventMax: number
+  ): SubEventFormItem[] {
+    const count = 2 + (seed % 3);
+    const totalMs = Math.max(2 * 60 * 60 * 1000, endMs - startMs);
+    const slotMs = Math.max(45 * 60 * 1000, Math.floor(totalMs / count));
+    const names = ['Kickoff', 'Main Session', 'Side Activity', 'Wrap-up'];
+    const items: SubEventFormItem[] = [];
+    for (let index = 0; index < count; index += 1) {
+      const optional = index > 0 && ((seed + index) % 2 === 0);
+      const stageStartMs = startMs + (index * slotMs);
+      const stageEndMs = index === count - 1 ? endMs : Math.min(endMs, stageStartMs + slotMs);
+      const slice = 0.45 + (((seed + index) % 4) * 0.12);
+      const capacityMax = Math.max(0, Math.round(eventMax * slice));
+      const capacityMin = optional ? 0 : Math.max(0, Math.min(capacityMax, Math.floor(capacityMax * 0.55)));
+      const accepted = Math.min(capacityMax, Math.max(0, Math.floor(capacityMin * 0.7)));
+      items.push({
+        id: `seed-${source.id}-casual-${index + 1}`,
+        name: `${names[index] ?? `Session ${index + 1}`}`,
+        description: `${source.shortDescription} (${index + 1}/${count})`,
+        startAt: AppUtils.toIsoDateTimeLocal(new Date(stageStartMs)),
+        endAt: AppUtils.toIsoDateTimeLocal(new Date(Math.max(stageStartMs + (30 * 60 * 1000), stageEndMs))),
+        createdByUserId: activeUserId,
+        groups: [],
+        optional,
+        capacityMin,
+        capacityMax,
+        membersAccepted: accepted,
+        membersPending: Math.max(0, capacityMax - accepted),
+        carsPending: (seed + index) % 3,
+        accommodationPending: (seed + index + 1) % 3,
+        suppliesPending: (seed + index + 2) % 4
+      });
+    }
+    return this.sortSubEventsByStartAsc(items);
+  }
+
+  private static buildSeededTournamentSubEvents(
+    source: EventMenuItem | HostingMenuItem,
+    startMs: number,
+    endMs: number,
+    seed: number,
+    activeUserId: string,
+    eventMax: number
+  ): SubEventFormItem[] {
+    const stageNames = ['Qualifiers', 'Semifinals', 'Finals'];
+    const stageCount = 3;
+    const totalMs = Math.max(3 * 60 * 60 * 1000, endMs - startMs);
+    const slotMs = Math.max(60 * 60 * 1000, Math.floor(totalMs / stageCount));
+    const items: SubEventFormItem[] = [];
+
+    for (let index = 0; index < stageCount; index += 1) {
+      const groupCount = Math.max(1, 4 >> index);
+      const basePerGroupMax = Math.max(2, Math.ceil(Math.max(2, eventMax) / Math.max(1, groupCount * (index + 1))));
+      const groups: SubEventGroupItem[] = [];
+      for (let groupIndex = 0; groupIndex < groupCount; groupIndex += 1) {
+        const groupMax = Math.max(2, basePerGroupMax - (groupIndex % 2));
+        const groupMin = Math.max(0, Math.floor(groupMax * 0.6));
+        groups.push({
+          id: `seed-${source.id}-s${index + 1}-g${groupIndex + 1}`,
+          name: `Group ${String.fromCharCode(65 + groupIndex)}`,
+          capacityMin: groupMin,
+          capacityMax: groupMax,
+          source: 'generated'
+        });
+      }
+      const totals = this.groupCapacityTotals(groups);
+      const stageStartMs = startMs + (index * slotMs);
+      const stageEndMs = index === stageCount - 1 ? endMs : Math.min(endMs, stageStartMs + slotMs);
+      const accepted = Math.min(totals.max, Math.max(0, Math.floor(totals.min * 0.7)));
+      items.push({
+        id: `seed-${source.id}-tournament-${index + 1}`,
+        name: `${stageNames[index]}`,
+        description: `${source.shortDescription} (${stageNames[index]})`,
+        startAt: AppUtils.toIsoDateTimeLocal(new Date(stageStartMs)),
+        endAt: AppUtils.toIsoDateTimeLocal(new Date(Math.max(stageStartMs + (45 * 60 * 1000), stageEndMs))),
+        createdByUserId: activeUserId,
+        groups,
+        tournamentGroupCount: groups.length,
+        tournamentGroupCapacityMin: Math.max(0, ...groups.map(group => Number(group.capacityMin) || 0)),
+        tournamentGroupCapacityMax: Math.max(0, ...groups.map(group => Number(group.capacityMax) || 0)),
+        tournamentLeaderboardType: (seed + index) % 2 === 0 ? 'Score' : 'Fifa',
+        tournamentAdvancePerGroup: index === stageCount - 1 ? 0 : Math.max(1, 2 - index),
+        optional: false,
+        capacityMin: totals.min,
+        capacityMax: totals.max,
+        membersAccepted: accepted,
+        membersPending: Math.max(0, totals.max - accepted),
+        carsPending: (seed + index) % 2,
+        accommodationPending: (seed + index + 1) % 2,
+        suppliesPending: (seed + index + 2) % 3
+      });
+    }
+    return this.sortSubEventsByStartAsc(items);
+  }
+
+  private static sortSubEventsByStartAsc(items: SubEventFormItem[]): SubEventFormItem[] {
+    return [...items].sort((a, b) => AppUtils.toSortableDate(a.startAt) - AppUtils.toSortableDate(b.startAt));
+  }
+
+  private static normalizedEventCapacityValue(value: number | null | undefined): number | null {
+    if (value === null || value === undefined) {
+      return null;
+    }
+    const parsed = Number(value);
+    if (!Number.isFinite(parsed)) {
+      return null;
+    }
+    return Math.max(0, Math.trunc(parsed));
+  }
+
+  private static groupCapacityTotals(groups: SubEventGroupItem[]): { min: number; max: number } {
+    let min = 0;
+    let max = 0;
+    for (const group of groups) {
+      const groupMin = Number(group.capacityMin);
+      const groupMax = Number(group.capacityMax);
+      const normalizedMin = Number.isFinite(groupMin) ? Math.max(0, Math.trunc(groupMin)) : 0;
+      const normalizedMax = Number.isFinite(groupMax) ? Math.max(normalizedMin, Math.trunc(groupMax)) : normalizedMin;
+      min += normalizedMin;
+      max += normalizedMax;
+    }
+    return { min, max };
+  }
+
+  private static eventStartAtMs(eventId: string, eventDatesById: Record<string, string>): number | null {
+    const iso = eventDatesById[eventId];
+    if (!iso) {
+      return null;
+    }
+    const value = new Date(iso).getTime();
+    return Number.isNaN(value) ? null : value;
+  }
+
+  private static eventFeedbackWhenLabel(eventId: string, eventDatesById: Record<string, string>): string {
+    const startMs = this.eventStartAtMs(eventId, eventDatesById);
+    if (startMs === null) {
+      return 'Recent event';
+    }
+    const parsed = new Date(startMs);
+    const day = parsed.toLocaleDateString('en-US', { weekday: 'short', month: 'short', day: 'numeric' });
+    const time = parsed.toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit' });
+    return `${day} · ${time}`;
+  }
+
+  private static feedbackHostUserForEvent(eventId: string, users: DemoUser[], activeUser: DemoUser): DemoUser {
+    const candidates = users.filter(user => user.id !== activeUser.id);
+    if (candidates.length === 0) {
+      return activeUser;
+    }
+    const index = this.hashText(`feedback-host:${eventId}`) % candidates.length;
+    return candidates[index] ?? candidates[0];
+  }
+
+  private static feedbackAttendeesForEvent(
+    eventId: string,
+    hostId: string,
+    users: DemoUser[],
+    activeUserId: string
+  ): DemoUser[] {
+    const candidates = users.filter(user => user.id !== activeUserId && user.id !== hostId);
+    if (candidates.length === 0) {
+      return [];
+    }
+    const seed = this.hashText(`feedback-attendees:${eventId}`);
+    const desired = Math.min(candidates.length, 3 + (seed % 4));
+    const picked: DemoUser[] = [];
+    for (let index = 0; index < candidates.length && picked.length < desired; index += 1) {
+      const candidate = candidates[(seed + (index * 3)) % candidates.length];
+      if (!candidate || candidate.id === activeUserId || picked.some(item => item.id === candidate.id)) {
+        continue;
+      }
+      picked.push(candidate);
+    }
+    return picked;
+  }
+
+  private static feedbackRoleForAttendee(eventId: string, attendeeUserId: string): 'Admin' | 'Manager' | 'Member' {
+    const seed = this.hashText(`feedback-role:${eventId}:${attendeeUserId}`);
+    if (seed % 11 === 0) {
+      return 'Admin';
+    }
+    if (seed % 4 === 0) {
+      return 'Manager';
+    }
+    return 'Member';
+  }
+
+  private static feedbackRoleToneClass(role: 'Admin' | 'Manager' | 'Member'): string {
+    if (role === 'Admin') {
+      return 'feedback-role-admin';
+    }
+    if (role === 'Manager') {
+      return 'feedback-role-manager';
+    }
+    return 'feedback-role-member';
+  }
+
+  private static feedbackRoleStatusClass(role: 'Admin' | 'Manager' | 'Member'): string {
+    if (role === 'Admin') {
+      return 'member-status-admin';
+    }
+    if (role === 'Manager') {
+      return 'member-status-manager';
+    }
+    return 'member-status-member';
+  }
+
+  private static feedbackRoleStatusIcon(role: 'Admin' | 'Manager' | 'Member'): string {
+    if (role === 'Admin') {
+      return 'admin_panel_settings';
+    }
+    if (role === 'Manager') {
+      return 'manage_accounts';
+    }
+    return 'person';
   }
 }
